Remove hardcoded `image_shape` from `config.py`
I previously pushed some updates that lets df3d determine the input
image size by just looking at the first image file, but I left the
default image size in config until being sure this worked. We've now run
df3d a lot on input images that have different sizes, and it works.
Running on different input sizes requires the user to _either_ update
the df3d.config['image_shape'] value for each dataset, or delete that
value which allows it to be automatically inferred. So let's just delete
the default value so that it can always automatically be inferred.